Changed FileMeshGenerator object to FileMesh object
The object that reads the mesh for the microreactors/gcmr/assembly/steady_state.bison_alone_transient_syntax
test is specified in the cli_args as a FileMesh. However, the input file contained
a FileMeshGenerator, which caused conflicts and ultimately, a test crash. This
commit changes the FileMeshGenerator in the input file to a FileMesh to be consistent
with what is provided in the test cli_args.

Ref #526